检索与知识层/Graphrag/requirements.txt
# ============================================================
# 检索与知识层 - GraphRAG 知识图谱构建
# ============================================================
# 功能：知识图谱构建、实体抽取、关系推理
# 环境：Python 3.11.13（调整为与项目统一）
# 更新时间：2025-10-21
# 状态：✅ 已验证（vLLM + Qwen2.5-7B）
# ============================================================

# ============================================================
# 1. 核心数据处理
# ============================================================
pandas==2.0.3
numpy==1.26.4
pyyaml==6.0.3
python-dotenv==1.1.1

# ============================================================
# 2. AI 服务客户端（vLLM OpenAI 兼容接口）
# ============================================================
openai==1.109.1
tiktoken==0.7.0
httpx==0.28.1
requests==2.32.5

# ============================================================
# 3. Neo4j 图数据库
# ============================================================
neo4j==5.25.0
# py2neo==2021.2.4  # 可选：高级接口

# ============================================================
# 4. 图算法和机器学习
# ============================================================
networkx==3.5
scikit-learn==1.3.2

# ============================================================
# 5. 文档处理
# ============================================================
# PyPDF2==3.0.1  # PDF解析（可选）
# python-docx==1.1.0  # Word解析（可选）
# markdown==3.5.0  # Markdown解析（可选）

# ============================================================
# 6. CSV 数据处理
# ============================================================
chardet==5.2.0

# ============================================================
# 7. 文件监控（可选）
# ============================================================
# watchdog==4.0.0

# ============================================================
# 8. 数据验证
# ============================================================
jsonschema==4.25.1
marshmallow==3.26.1

# ============================================================
# 9. 测试框架（开发环境）
# ============================================================
pytest==7.4.4
pytest-asyncio==0.23.6
pytest-cov==4.1.0

# ============================================================
# 10. 日志工具
# ============================================================
# structlog==24.0.0  # 结构化日志（可选）
colorama==0.4.6
loguru==0.7.2

# ============================================================
# 11. 代码质量工具（开发环境）
# ============================================================
black==23.12.1
flake8==6.1.0
# mypy==1.8.0  # 类型检查（可选）

# ============================================================
# 12. 异步支持
# ============================================================
aiofiles==23.2.1

# ============================================================
# 13. 进度显示
# ============================================================
tqdm==4.66.3

# ============================================================
# vLLM 服务端说明（单独环境）
# ============================================================
# vLLM 服务端建议在单独环境中安装：
# 
# conda create -n vllm python=3.11
# conda activate vllm
# pip install torch==2.1.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# pip install vllm
# 
# 启动服务：
# python -m vllm.entrypoints.openai.api_server \
#   --model /path/to/Qwen2.5-7B-Instruct \
#   --served-model-name qwen2.5-7b \
#   --port 8000
# ============================================================

# ============================================================
# 使用说明
# ============================================================
# 
# 安装：
# pip install -r requirements.txt
# 
# 运行图谱构建：
# python src/build_kg.py
# 
# 注意：
# - vLLM 需要单独环境和服务
# - Neo4j 需要单独启动
# - 神农数据集需要预处理
# ============================================================
